First version of the usb protocol
after reading the wikipedia article on HDLC I was inspried to created a similar protocol for thermodroid.

the first 4 bytes of every frame are referred to as a Header.  Header byte order is as follows:

byte 1 : Frame Delimiter
byte 2 : Command (state)
byte 3 and 4 : the length of the data frame ( without the 4 header bytes)